Restructure compliance engine to have core and engine. Add queue for storing entities. Implement processing logic. Refactoring
